requirements.txt
agate==1.9.1
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
comm==0.2.3
daff==1.4.2
dbt-adapters==1.16.7
dbt-common==1.32.0
dbt-core==1.10.13
dbt-extractor==0.6.0
dbt-postgres==1.9.1
dbt-protos==1.0.375
dbt-semantic-interfaces==0.9.0
debugpy==1.8.17
decorator==5.2.1
deepdiff==8.6.1
executing==2.2.1
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
isodate==0.6.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
leather==0.4.0
MarkupSafe==3.0.3
mashumaro==3.14
matplotlib-inline==0.1.7
more-itertools==10.8.0
msgpack==1.1.1
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.3
orderly-set==5.5.0
packaging==25.0
pandas==2.3.2
parsedatetime==2.6
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
protobuf==6.32.1
psutil==7.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
snowplow-tracker==1.1.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
stack-data==0.6.3
text-unidecode==1.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
zipp==3.23.0
